requirements.txt
streamlit
pandas
openpyxl
python-barcode
python-dotenv
extra-streamlit-components
pillow